@@ -1,10 +1,10 @@
-import {Response} from '@angular/http';
+import {HttpErrorResponse} from '@angular/common/http';
 import 'rxjs/add/observable/throw';
 import { Observable } from 'rxjs/Observable';
 
 export class ErrorHandler {
 
-    static handleError(error: Response | any) {
+    static handleError(error: HttpErrorResponse | any) {
         let errorMessage: string;
 
         if (error instanceof Response) {
