@@ -3,13 +3,13 @@ import { ShoppingCartService } from 'app/restaurant-detail/shopping-cart/shoppin
 import { CartItem } from 'app/restaurant-detail/shopping-cart/cart-item.model';
 import { Order } from './order.model';
 import { Observable } from 'rxjs/Observable';
-import { Http, RequestOptions, Headers } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { MEAT_API } from 'app/app.api';
 
 @Injectable()
 export class OrderService {
 
-    constructor(private cartService: ShoppingCartService, private http: Http) {
+    constructor(private cartService: ShoppingCartService, private http: HttpClient) {
 
     }
 
@@ -34,14 +34,8 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> {
-        const headers = new Headers();
-        headers.append('Content-Type', 'application/json');
-        return this.http.post(`${MEAT_API}/orders`,
-                                JSON.stringify(order),
-                                new RequestOptions({headers: headers}))
-                                .map(response => response.json())
-                                // tslint:disable-next-line:no-shadowed-variable
-                                .map(order => order.id);
+        // tslint:disable-next-line:no-shadowed-variable
+        return this.http.post<Order>(`${MEAT_API}/orders`, order).map(order => order.id);
     }
 
     clear() {
